Kickstart updates to resolve prepatching issues
The prepatching feature changed how kickstarts are delivered. Where
before, they were generated as part of build-iso and copied during
installation, they are now delivered in packages. However, when the
controller kickstart mirrors the feed directory from the active
controller after installation, it may overwrite these installed files.
This update changes the wget command options to protect against this.

In addition, testing with an RT kernel patch showed Anaconda was also
installing the mlnx-ofa_kernel-rt-modules package from the patching
repo on a standard note, as it attempts to resolve a packaging
requirement. This update also adds explicit exclusions to the package
lists in the standard and lowlatency kickstarts to avoid installing
rt modules on the standard nodes, and vice versa.

Change-Id: I56b22fb0846db05a96004184c1060c05566d5363
Signed-off-by: Don Penney <don.penney@windriver.com>