stx-metal: API ref doc content added.
I added the api-ref-sysinv-v1-metal.rst file, which is content
for the API reference manual to the api-ref/source dir. This
represents the converted old-style files to the newer OpenStack
supported RST files.

I updated the index.rst file to include the new .rst file so that
the api-ref document can build.

Updated the actual output using Platform http://10.10.10.2:6385/v1
and querying.

Change-Id: I943dd136b2b083743e053db5070d4da7fbe78685
Signed-off-by: Scott Rifenbark <srifenbark@gmail.com>