Fix bug with PXE Boot Server
The UEFI grub config file used for PXE Boot Server containes a mixed
of menuentry  and submenu. A submenu opens a new context and global
variables cannot be used within them
https://bugs.launchpad.net/ubuntu/+source/grub2/+bug/1175127. This
was causing pxe_root to be empty for submenus selections. Moving
'1) UEFI Boot from hard drive' into a submenu fixes this problem and
also removes the need to save the root variable. This was previously
required when a user would select boot from hard drive and the hard
drive did not exist, root would no longer be set to the pxeboot
server and selecting the other menu entries would fail.

Remove the '1)' prefix in the 'Boot from hard drive' menu option in
both Legacy and UEFI boot as this is not used by other menues.

Closes-Bug: 1794863

Change-Id: I5bc62039bfb68477e9cd0166ce17877693037640
Signed-off-by: Kristine Bujold <kristine.bujold@windriver.com>