Add mtce dependency on ipmitool
ipmitool was recently found to be missing from the load after
a rpm cleanup that seemed to remove all dependency on it.

Maintenance and its Hardware Monitor use the ipmitool
for power / reset control as well as sensor monitoring.

This update adds a dependency on ipmitool in the maintenance
mtcAgent and hwmon rpm build recipe so that it will always
be included in the load with maintenance.

Closes-Bug:1821958

Test Plan:
PASS: Verify ipmitool in load
PASS: Verify mtce and hwmon rpm dependency on ipmitool
PASS: Verify system install

Change-Id: I958a2365f6df7bdbf942bc57c1aa17ee2ae6a73d
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>