Add mtce support for manifest apply over initial controller unlock
The introduction of Ansible requires the execution of a manifest
as part of the first controller's initial unlock.

Unfortunately maintenance issues the lazy self reboot immediately
upon receiving the unlock command, interrupting the in-progress
manifest apply.

This update identifies the initial self reboot of the only
provisioned host condition and waits for up to a timeout
period for an unlock ready signal that is provided by
successful completion of the 'initial-unlock-manfest'.

Seeing the unlock ready signal prior to the timeout allows
the unlock self reboot to proceed normally.

Depends-On:https://review.openstack.org/#/c/643914
Story:2004695
Task:30243

Test Plan:
PASS: Verify timeout handling - allowing retry
PASS: Verify with signal - immediate
PASS: Verify with signal - before timeout

Change-Id: I3633e772310c36af5df57364f66c14f037b2ea8f
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>