Exclude SRIOV i40evf-rt, ixgbevf-rt modules
To enable the SRIOV device plugin feature on a low latency system,
the i40evf-rt and ixgbevf-rt drivers need to be included in the
host image.

However, we want to ensure these modules are not pulled into a non
low-latency system.

As such, these modules have been added to the non low-latency
filter lists.

Closes-Bug: #1828848

Change-Id: Ic7b134d72bf19c19845e86798b36bccd40e0bbff
Signed-off-by: Steven Webster <steven.webster@windriver.com>