Remove sdk package from python-inventoryclient.spec
A previous update removed the "%files sdk" section from
the python-inventoryclient.spec file, but missed removal
of the "%package sdk" section. As a result, the build
system looks for the python-inventoryclient-sdk package
when checking if a rebuild is required. But because the
"%files" was removed, the package is not built. This leads
the build system to decide it needs to rebuild.

Change-Id: I753212a17eca59699c2122c55eac313d1f9ed94f
Partial-Bug: 1830284
Signed-off-by: Don Penney <don.penney@windriver.com>