B&R: Skip Ceph osds from getting wiped during initial installation
To keep ceph-mon data intact at restore, ceph osds should not be wiped.
Every ceph osd has a Globally unique identifier(GUID) and based on this
GUID ceph osds are not added in list of devices to wipe.
Since installation is the same on restore as on a fresh install we need
to keep ceph data intact then delegate the wipe to ansible if we are not
at restore (or we are at restore but user does not want to keep ceph's
data). To keep the previous behavior, when any other node is installed
we wipe Ceph.

Story: 2004761
Task: 34283
Change-Id: Ief93da56c1344ab259e674ca8d79f9952fefcd2b
Signed-off-by: Elena Taivan <elena.taivan@windriver.com>