Enhance filtering of software groups for install
Software groups are generated using the platform_comps.py
utility to declare what software is installed on the
various node types, generating a comps.xml for the base
software repo in the ISO (and on the controller) with
package groups for each node type. The "filter_out_from_X"
files provide a list of packages to exclude from the
corresponding software group.

This update enhances the filtering by using the package
list from the image.inc files, rather than the full set of
packages included in the ISO. The latter, which is the existing
implementation, would include all packages pulled in to meet
dependency requirements in addition to those specified in
image.inc files. As a result, we would also need to include
these dependencies in the filter files to ensure they are not
installed on a given node.

By using just the package list from the image.inc files, the
software groups will only include the requested packages. Adding
a package to the filter file would then effectively also filter
its dependencies.

For example, if package A is in an image.inc file and requires
package B, and we include A in the filter_out_from_storage list,
the existing implementation would still include B in the storage
software group and B would be installed on the node. With this
updated implementation, B would not be included in the software
groups at all, and only installed on nodes that also install A.

Additionally, this update also cleans up filter references from
centos-ks-gen.pl which were obsolete and should have been removed
previously.

This update also adds starlingx-dashboard to the filter lists for
storage and worker nodes.

As an example, this update has the net effect of reducing the
number of packages installed on a worker node from the current
total of 1060 to 932 packages, by trimming the unneeded packages
pulled in from dependencies of filtered packages.

Change-Id: I1c1c3fabf53c4a1dd085f135c7443e09f96e5906
Story: 2004764
Task: 33602
Signed-off-by: Don Penney <don.penney@windriver.com>