Supress implicit-fallthrough warnings.
In some parts of the mtce code some implicit fallthrough are used.
This causes a warning in the compiler and in OSes like openSUSE the
-Werror flag is enforced leading to a build error.

In this commit the __attribute__ ((fallthrough)) is used to tell the
compiler to not worry about this implicit fallthroughs as the it works
as intended.

Change-Id: I219ff9d490f3f86ad045e0f0e891f40467baaf06
Story: 2005862
Task: 33667
Signed-off-by: Erich Cordoba <erich.cordoba.malibran@intel.com>