Change enabled_count variable from bool to int in nodeLinkClass
The enabled_count member inside nodeLinkClass has a declared bool
datatype, but this variable is used as a counter, where an int data
type is a better choice. This change updates the datatype to int.

Also, there was a misleading-indentation warning in mtcNodeHdlrs.cpp,
which is also fixed in this change.

Change-Id: Ib154c5b6ae2e7068870733b5ee8971e20cedb43f
Story: 2005862
Task: 34163
Signed-off-by: Erich Cordoba <erich.cordoba.malibran@intel.com>