Cleanup unused constants
As part of the cleanup tasks for openstack, there are a number
of attributes that are being cleaned up such as concurrent_disk_operations
and instance_backing.

The inventory repo has duplicated references to these, so many of them
are being removed here to coincide with the cleanup in sysinv.

Note: there are far more values that can be cleaned up, but this is mainly
to remove all traces of the deprecated features, rather than a generic
cleanup of inventory.

Change-Id: I6d582c83d6e86aa6a08cb110e9f009ff9549006d
Story: 2004764
Task: 34274
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>