Fix format-overflow warning in mtcInvApi
There is a format-overflow warning because the batch_str buffer
is 10 chars length and the formatted variable can get at most 11 chars.

To fix this issue the sprintf was changed to snprintf to avoid the
overflow. Also the signature of mtcInvApi_read_inventory was changed
to avoid negative values.

The impact of this change is minimum as this function only has one
caller, where the supplied argument is always the number 5.

Change-Id: I3adfd270e603021d6e41bdad08eb06e65686b989
Signed-off-by: Erich Cordoba <erich.cordoba.malibran@intel.com>