Only install pm-qos-mgr on worker and AIO nodes
The pm-qos-mgr package should only be installed on worker nodes.
This package needs to be excluded from controllers.
It had only been excluded from storage.
This is not a functional problem, but impacted patching
on controllers if we ever have to update it.

pm-qos-mgr should match the packaging (filters) of worker-utils.

Change-Id: I7a68d01be1e4d7dd6f3ef327ccce795362643515
Partial-Bug: 1840356
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>