Allow computes without virtualization enabled to be unlocked
The goenabled script in mtce-compute checks if hardware
virtualization is enabled, before allowing the node to
become enabled.  (hardware only)

This check is only needed if the compute is being used as an
openstack compute node, and so the check has been updated
to look for the appropriate label.

Puppet stores the labels in the hieradata file in /tmp.

That file is recreated when the host is rebooted or unlocked,
and labels cannot be changed unless a host is locked.

Change-Id: I6862a1b7824c759f0c3caea37c3ae6e4e9ab43b7
Closes-Bug: 1843616
Signed-off-by: Al Bailey <Al.Bailey@windriver.com>