Fix rare mtcAgent segfault on process shutdown
A mtcAgent core dump was observed during unit testing
of another feature. Debug of that coredump revealed the
segfault occured in the freeing of host object memory.

Occurrence is extremely rare but at the same time there
is no real need to free this memory in the destructor
because the kernel, which is much better at that task,
does that automatically when a proccess exits.

Test Plan:

PASS: Verify system install with no core dumps.
PASS: Verify mtcAgent proccesss restart soak with no code dumps.

Change-Id: I6107078fb802be0ef2aaf632b79e751376ba9c42
Closes-Bug: 1831956
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>