Fix Mtce's VIM systems query handling
A bug was recently introduced by improperly trying to fix
the handling of an invalid host lookup error in the mtce
http server handler.

The impact of this issue is to prolong the VIM DOR mode
to the point of its timeout.

This update makes the necessary changes to remove the new
issue and properly handle the original issue.

This update also improves client code handling and ensures the
service url length is loaded correctly based on client making
the request.

Test Plan:

PASS: Verify AIO DX System Install
PASS: Verify maintenance logs for evidence of corrected behavior

Change-Id: I0fe09b5a17f5fc1a6dd84876cef575dd15c88db2
Closes-Bug: 1847334
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>