Set IPV6_AUTOCONF=no on initial onboot devices
This commit sets IPV6_AUTOCONF=no from the anaconda installer for
all interfaces that are enabled during the initial install.

Once the puppet manifest is applied, and it updates the networking
config scripts for the system, this will be overwritten by that config,
which disables it for all configured interfaces,

IPv6 auto-configuration is causing an issue during backup and restore,
where router advertisements are causing the default route to be deleted,
which causes the calico pod to fail to initialize.

Closes-bug: 1845707
Change-Id: I519ad0d92c66a636df0d10e79c6962296683520d
Signed-off-by: Joseph Richard <joseph.richard@windriver.com>