Update host watchdog CONFIG_MASK
The CONFIG_MASK in hostw.h includes CONFIG_START_DELAY, which
correlates to an option that is not actually used in host watchdog. As
a result, the recently added check that verifies all options in the
CONFIG_MASK are set fails, and the host watchdog fails to launch.

This update removes the unused CONFIG_START_DELAY bit from
CONFIG_MASK.

Change-Id: I330e15520bc0f01a6cbfd4f83a1953c1c737da2b
Partial-Bug: 1835370
Signed-off-by: Don Penney <don.penney@windriver.com>