Update kickstarts to remove cgts-vg as needed
In some scenarious, such as when the cgts-vg volume group spans
multiple disks, the disk wiping in the kickstarts is insufficient to
clear existing LVM metadata, and anaconda reports a failure to create
the cgts-vg volume group as a result.

This update adds explicit cleanup of the logical and physical volumes
associated with cgts-vg, if not already wiped by the user prior to
installation, and removes cgts-vg itself. This is done ahead of the
anaconda partitioning, avoiding any conflict with automatic creation
of the new cgts-vg.

Change-Id: Id33cab02e1cecbfee8035cc55213c99ef3cea34d
Closes-Bug: 1854082
Signed-off-by: Don Penney <don.penney@windriver.com>