Update pxeboot kickstart to allow for hybrid install
The generated pxeboot kickstart templates that are included in the
installation ISO will be used by the gen-bootloader-iso.sh in setting
up a hybrid ISO/network installation, where the kernel and initrd are
loaded from ISO, but the rootfs, kickstart, and software repositories
are accessed via http/https. In order to support this installation,
the pxeboot kickstart is updated to allow for the case where the
BOOTIF cmdline parameter is not set (ie. node booted from ISO rather
than network).

This update also adds a check to the pxeboot kickstart to mount the
installation media if in this scenario. In the case where a controller
has been booted from an ISO with installation media accessed from the
network, dracut does not mount the ISO. This results in the 'reboot
--eject' in the kickstart being ignored, and when the newly installed
node is rebooted, it may just boot the ISO again. By mounting the
media in this case, Anaconda will eject it as expected.

Story: 2006980
Task: 38466

Change-Id: I9330a2d50b8966a2b3f2b59cabec20f084372d8c
Signed-off-by: Don Penney <don.penney@windriver.com>