Select disk via kickstart
If the boot_device and/or rootfs_device are not set on the kernel
cmdline detect them early in the kickstart script. This can help
solve the issue about which disk type is to be used.

Remove the older code for disk detection
Tested with both sda and nvme disk types

Story: 2007486
Task: 39204
Change-Id: I4fa3b44a4e656e280820ceeefafaf127cb048df6
Signed-off-by: Saul Wold <sgw@linux.intel.com>