Drop copy of .cfg files from controller kickstarts
In a boot from an ISO modified by update-iso.sh with a ks-addon, the
ks-addon.cfg was being unnecessarily copied in a kickstart postinstall
scriptlet. This behaviour, copying the .cfg files, was originally
required for copying the net-boot kickstarts for installation of other
nodes, however the kickstarts are now packaged and installed, and no
other .cfg files are required from the installation.

As such, this update drops the copying of these files.

Change-Id: Id088ff00711b0703299f822ab1f25901e94a6d4d
Closes-Bug: 1875464
Signed-off-by: Don Penney <don.penney@windriver.com>