Fix partition removal after wipe
After the wipe step, partitions are not removed when installing the load.
This commit fixes this.

Also, on some systems with NVMe, udev doesn't correctly remove the device
nodes to the deleted partitions from /dev/nvme* causing them to be seen as
non block devices, this leads to failures on format or assigning LVM PVs.

Change-Id: I3ab9f70d00a38568fc00063cdaa54ec3be48dc58
Closes-Bug: 1876374
Signed-off-by: Ovidiu Poncea <ovidiu.poncea@windriver.com>