Wipe OSD journals during host reinstall
A host reinstall (using the system host-reinstall command) fails
if the OSD journals are not wiped before reinstalling.

Based on the ceph-manage-journal.py script, just wiping the
standard 17KB at the beginning and end of the journal partition
is not enough, but instead about 100MB of data must be wiped
instead.

Change-Id: I165c385958f7f700cae28312998276aa69ed22c3
Closes-bug: 1860165
Signed-off-by: Stefan Dinescu <stefan.dinescu@windriver.com>