Fix worker/storage disk setup in kickstart files
In Anaconda, all the disk and partitions instructions
are present in the /tmp/part-include file.
In the storage and compute kickstart files,
the /tmp/part-include file is truncated
when an upgrade is triggered.
We fix this by changing "cat<<EOF>/tmp/part-include"
to "cat<<EOF>>/tmp/part-include" in the kickstart files.

Closes-Bug: 1881776
Change-Id: Ic3818127ac7477d534156ed28c16cb11123f0424
Signed-off-by: Mihnea Saracin <Mihnea.Saracin@windriver.com>