Enhance kickstart check for backup partition
Add a check of the FS type to the kickstart when checking for a
pre-existing backup partition. If a node has been installed with
starlingx using EFI, setting up the first partition to be the backup
partition, and then reinstalled with a non-starlingx OS that uses
the first partition for EFI boot, there is a possiblity that the GUID
metadata is not reset by the non-starlingx OS. When the node is
reinstalled with starlingx, checking just the GUID could make it
appear this EFI boot partition is actually a pre-existing backup.

Enhancing this check to also validate that the FS type is ext4 ensures
that the backup partition at least has the correct FS type and can be
resized by the kickstart if needed.

Change-Id: Ibb64efb495bd07b7999d068f2e3b2dc37fd9cae4
Close-Bug: 1887448
Signed-off-by: Don Penney <don.penney@windriver.com>