Stop the process monitor (pmond) on controlled self-reboot
There are still cases seen where on an AIO SX unlock operation
fails to reboot due to pmond recovering the mtcClient following
a mtcClient self-reboot and launch of fail-safe sysreq reset thread.

Following a self-reboot, the Process Monitor (pmond) detects an active
monitoring failure of the mtcClient. However, at that same time
systemctl reports that the system is running degraded, not stopping.

So the previous fix to pmon does not know that the system is stopping
so it restarts mtcClient ; like before but valid systemctl state
readout.

This update is a further enhancement for the issue reported by
https://bugs.launchpad.net/starlingx/+bug/1883519 with update
https://review.opendev.org/#/c/735609 by commanding the mtcClient
to stop pmond, with verification and retries, immediately before
a self-reboot.

Change-Id: I17fde797803c537f4f448b4764585f1f1acc4e2a
Closes-Bug: 1883519
Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>