Align partitions created by kickstarters
Partitions on some disks may be created unaligned.

The cause is that the creation of partitions is done between
specific intervals expressed in MBs. The kernel exposed a
specific variable for each disk for providing an offset to
align each partitions (/sys/block/<disk>/alignment_offset).

For better granular control, we transform MB units into
logical sector units and use the alignment_offset variable
to properly align the partitions.

Change-Id: I971c232fe0969eac14b85c5796908f0c85e23dbf
Closes-bug: 1883975
Signed-off-by: Daniel Safta <daniel.safta@windriver.com>